Add Qa app

+++ yuno_sdk_android/android/src/main/kotlin/com/yuno_flutter/yuno_sdk_android/features/payment_methods/views/YunoPaymentMethodView.kt
@@ -52,7 +52,7 @@ class YunoPaymentMethodView(
             if (currentHeightPx <= 0) return@OnGlobalLayoutListener
 
             val density = context.resources.displayMetrics.density
-            val currentHeightDp = currentHeightPx / density
+            val currentHeightDp = (currentHeightPx / density).toDouble()
 
             // Avoid spamming the channel for tiny diffs during animation/layout passes.
             if (abs(currentHeightDp - lastHeightDp) > 0.5) {
